+++ dinky-web/src/services/constants.tsx
@@ -353,27 +353,3 @@ export const DIALECT = {
   SQL: "sql",
   JAVASCRIPT: "javascript",
 };
-
-export const JOB_LIFE_CYCLE = {
-  UNKNOWN: 0,
-  CREATE: 1,
-  DEVELOP: 2,
-  DEBUG: 3,
-  RELEASE: 4,
-  ONLINE: 5,
-  CANCEL: 6,
-};
-
-export const JOB_STATUS = {
-  FINISHED: 'FINISHED',
-  RUNNING: 'RUNNING',
-  FAILED: 'FAILED',
-  CANCELED: 'CANCELED',
-  INITIALIZING: 'INITIALIZING',
-  RESTARTING: 'RESTARTING',
-  CREATED: 'CREATED',
-  FAILING: 'FAILING',
-  SUSPENDED: 'SUSPENDED',
-  CANCELLING: 'CANCELLING',
-  UNKNOWN: 'UNKNOWN',
-};
